+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Allows to disable coercion
+
 ## 2.0.0-alpha.4 - 2020-03-21
 
 - Publish only src folder for jest-structure